+++ src/partialMatchRegExp/constants.ts
@@ -1,3 +1,4 @@
 export const NOT_NUMBERS_REGEX = /\D/g;
 export const FLAGS_IRRELEVANT_TO_REBUILD = /[dgy]/g;
+export const LITERAL_BACKSLASH = "\\\\";
 export const LITERAL_K = "k";